requirements.txt
fastapi
streamlit
langchain
openai
transformers
sentence-transformers
pydantic
pymarkdown
uvicorn
docker
numpy 
pandas
langchain-community
chromadb
faiss-cpu
google-generativeai
python-dotenv
black
isort
torch